+++ ReleaseNotes.md
@@ -9,6 +9,8 @@ Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in 
 New analyzers are considered "minor" changes (even though adding a new analyzer is likely to generate warnings
 or errors for existing code when the package is upgraded).
 
+* `[minor]` Fix FL0024 code fix provider to properly find interpolated string expressions within argument nodes.
+
 ## 1.6.0 Beta 2
 
 * Add [FL0023](https://github.com/Faithlife/FaithlifeAnalyzers/wiki/FL0023): Avoid obsolete `Logos.Common.Logging.Extensions.LoggerExtensions` methods.